[server] Separate learn contract related code from class API
